added topArtist's shimmer effect loading
